requirements.txt
flask==3.0.2
flask-cors==4.0.0
groq==0.4.2
gunicorn==21.2.0
python-dotenv==1.0.1 